feat(nvim_ufo): improve fold end line detection logic
Refactors `nvim_ufo` configuration to use a new `contain_end_line`
function, which leverages a filetype-specific table `ft_end` to
determine when to display the end line of a folded block.

This change replaces a large `if-elseif` block with a more organized and
extensible approach, improving readability and maintainability. It adds
specific rules for `html`, `javascript`, `lua`, `markdown`, `python`,
`ruby`, `sh`, and `vim` filetypes, and also includes a fallback for
other types.

Additionally, `css`, `html`, and `javascript` are now explicitly set to
use the `treesitter` fold provider, and `http` is added to
`close_fold_kinds_for_ft`. Preview mappings for scrolling and jumping
are also introduced.