feat: add nixd LSP and Nixfmt formatter
Adds `nixd` as the LSP for Nix files, replacing the previous `rnix-lsp`
setup.
Also switches the Nix formatter from `nixpkgs_fmt` to `nixfmt` in
conform.

A new check is added to the format keymap to prevent formatting when the
buffer is not modifiable.
The diagnostic virtual text prefix is updated from `U+F4C3` to `•`.